LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpParameters-maxFramerate-expected.txt


FAIL addTransceiver() with sendEncoding.maxFramerate field set to less than 0 should reject with RangeError assert_throws_js: function "() => pc.addTransceiver('video', {
    sendEncodings: [{
      maxFramerate: -10
    }]
  })" did not throw
PASS addTransceiver('audio') with sendEncoding.maxFramerate should succeed, but remove the maxFramerate, even if it is invalid
FAIL setParameters with maxFramerate on an audio sender should succeed, but remove the maxFramerate assert_not_own_property: unexpected property "maxFramerate" is found on object
FAIL setParameters with an invalid maxFramerate on an audio sender should succeed, but remove the maxFramerate assert_not_own_property: unexpected property "maxFramerate" is found on object
FAIL setParameters() with encoding.maxFramerate field set to less than 0 should reject with RangeError assert_unreached: Should have rejected: undefined Reached unreachable code
PASS setParameters() with maxFramerate 24->16 should succeed with RTCRtpTransceiverInit
PASS setParameters() with maxFramerate 24->16 should succeed without RTCRtpTransceiverInit
PASS setParameters() with maxFramerate undefined->16 should succeed with RTCRtpTransceiverInit
PASS setParameters() with maxFramerate undefined->16 should succeed without RTCRtpTransceiverInit
FAIL setParameters() with maxFramerate 24->undefined should succeed with RTCRtpTransceiverInit assert_equals: expected (undefined) undefined but got (number) 24
FAIL setParameters() with maxFramerate 24->undefined should succeed without RTCRtpTransceiverInit assert_equals: expected (undefined) undefined but got (number) 24
FAIL setParameters() with maxFramerate 0->16 should succeed with RTCRtpTransceiverInit promise_test: Unhandled rejection with value: object "RangeError: maxFrameRate is below or equal 0"
PASS setParameters() with maxFramerate 0->16 should succeed without RTCRtpTransceiverInit
PASS setParameters() with maxFramerate 24->0 should succeed with RTCRtpTransceiverInit
PASS setParameters() with maxFramerate 24->0 should succeed without RTCRtpTransceiverInit
